Task 288, added logo and made a separate folder for images and updated the import image statements for few files.
